Add files via upload
The bot works by first using the Wikipedia API to retrieve the text of the article that the user has asked about. It then passes this text to the LLM, which generates a response. The response is then filtered and edited by the bot to ensure that it is accurate and relevant.